feat: a hook for `dev/bin` utilities
the `dev` repository contains a `bin/` directory with various justfiles
referred to by the linkerd2-proxy's own justfile.

for example, building a proxy image via `just docker` will fail without
this in the path because it depends on `just-cargo`.

this commit introduces a conditional expression so that the `bin/`
directory is added to the path if an environment variable exists.

this shell can now be called by running a command like:

```
LINKERD_DEV="/some/path" nix-shell shell.nix
```

Signed-off-by: katelyn martin <git@katelyn.world>